backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<3.16
django-filter>=24.2
drf-spectacular>=0.27
django-cors-headers>=4.3
django-environ>=0.11
psycopg2-binary>=2.9

